+++ src/main/java/slimeknights/toolleveling/ToolLevelNBT.java
@@ -10,7 +10,7 @@ public class ToolLevelNBT extends ModifierNBT {
   private static final String TAG_XP = "xp";
   private static final String TAG_BONUS_MODIFIERS = "bonus_modifiers";
 
-  public long xp; // current level xp
+  public int xp; // current level xp
   public int bonusModifiers; // saved extra so tools don't change unexpectedly on config change
 
   public ToolLevelNBT(NBTTagCompound tag) {
@@ -20,14 +20,14 @@ public class ToolLevelNBT extends ModifierNBT {
   @Override
   public void read(NBTTagCompound tag) {
     super.read(tag);
-    xp = tag.getLong(TAG_XP);
+    xp = tag.getInteger(TAG_XP);
     bonusModifiers = tag.getInteger(TAG_BONUS_MODIFIERS);
   }
 
   @Override
   public void write(NBTTagCompound tag) {
     super.write(tag);
-    tag.setLong(TAG_XP, xp);
+    tag.setInteger(TAG_XP, xp);
     tag.setInteger(TAG_BONUS_MODIFIERS, bonusModifiers);
   }
